app: Move out of enterprise and merge with cmd/sourcegraph-oss
Another item on the checklist to get rid of the enterprise directory.
