+++ cmd/blobstore/main.go
@@ -3,7 +3,7 @@ package main // import "github.com/sourcegraph/sourcegraph/cmd/blobstore"
 
 import (
 	"github.com/sourcegraph/sourcegraph/cmd/blobstore/shared"
-	"github.com/sourcegraph/sourcegraph/cmd/sourcegraph-oss/osscmd"
+	"github.com/sourcegraph/sourcegraph/cmd/sourcegraph/osscmd"
 	"github.com/sourcegraph/sourcegraph/internal/sanitycheck"
 )
