+++ cmd/frontend/BUILD.bazel
@@ -10,7 +10,7 @@ go_library(
     visibility = ["//visibility:private"],
     deps = [
         "//cmd/frontend/shared",
-        "//cmd/sourcegraph-oss/osscmd",
+        "//cmd/sourcegraph/osscmd",
         "//internal/sanitycheck",
         "//ui/assets",
         "//ui/assets/oss",
