+++ cmd/github-proxy/BUILD.bazel
@@ -11,7 +11,7 @@ go_library(
     visibility = ["//visibility:private"],
     deps = [
         "//cmd/github-proxy/shared",
-        "//cmd/sourcegraph-oss/osscmd",
+        "//cmd/sourcegraph/osscmd",
         "//internal/sanitycheck",
     ],
 )
