+++ cmd/github-proxy/github-proxy.go
@@ -2,7 +2,7 @@ package main
 
 import (
 	"github.com/sourcegraph/sourcegraph/cmd/github-proxy/shared"
-	"github.com/sourcegraph/sourcegraph/cmd/sourcegraph-oss/osscmd"
+	"github.com/sourcegraph/sourcegraph/cmd/sourcegraph/osscmd"
 	"github.com/sourcegraph/sourcegraph/internal/sanitycheck"
 )
