+++ cmd/gitserver/BUILD.bazel
@@ -11,7 +11,7 @@ go_library(
     visibility = ["//visibility:private"],
     deps = [
         "//cmd/gitserver/shared",
-        "//cmd/sourcegraph-oss/osscmd",
+        "//cmd/sourcegraph/osscmd",
         "//internal/sanitycheck",
     ],
 )
