--- cmd/sourcegraph-oss/BUILD.bazel
@@ -1,32 +1,0 @@
-load("@io_bazel_rules_go//go:def.bzl", "go_binary", "go_library")
-
-go_library(
-    name = "sourcegraph-oss_lib",
-    srcs = ["main.go"],
-    importpath = "github.com/sourcegraph/sourcegraph/cmd/sourcegraph-oss",
-    visibility = ["//visibility:private"],
-    deps = [
-        "//cmd/blobstore/shared",
-        "//cmd/frontend/shared",
-        "//cmd/github-proxy/shared",
-        "//cmd/gitserver/shared",
-        "//cmd/repo-updater/shared",
-        "//cmd/searcher/shared",
-        "//cmd/sourcegraph-oss/osscmd",
-        "//cmd/symbols/shared",
-        "//cmd/worker/shared",
-        "//internal/sanitycheck",
-        "//internal/service",
-        "//internal/service/servegit",
-    ],
-)
-
-go_binary(
-    name = "sourcegraph-oss",
-    embed = [":sourcegraph-oss_lib"],
-    visibility = ["//visibility:public"],
-    x_defs = {
-        "github.com/sourcegraph/sourcegraph/internal/version.version": "{STABLE_VERSION}",
-        "github.com/sourcegraph/sourcegraph/internal/version.timestamp": "{VERSION_TIMESTAMP}",
-    },
-)
