--- cmd/sourcegraph-oss/osscmd/BUILD.bazel
@@ -1,13 +1,0 @@
-load("@io_bazel_rules_go//go:def.bzl", "go_library")
-
-go_library(
-    name = "osscmd",
-    srcs = ["osscmd.go"],
-    importpath = "github.com/sourcegraph/sourcegraph/cmd/sourcegraph-oss/osscmd",
-    visibility = ["//visibility:public"],
-    deps = [
-        "//internal/authz",
-        "//internal/service",
-        "//internal/service/svcmain",
-    ],
-)
