+++ cmd/sourcegraph/BUILD.bazel
@@ -3,7 +3,7 @@ load("@io_bazel_rules_go//go:def.bzl", "go_binary", "go_library")
 go_library(
     name = "sourcegraph_lib",
     srcs = ["main.go"],
-    importpath = "github.com/sourcegraph/sourcegraph/enterprise/cmd/sourcegraph",
+    importpath = "github.com/sourcegraph/sourcegraph/cmd/sourcegraph",
     visibility = ["//visibility:private"],
     deps = [
         "//cmd/blobstore/shared",
@@ -13,9 +13,9 @@ go_library(
         "//cmd/precise-code-intel-worker/shared",
         "//cmd/repo-updater/shared",
         "//cmd/searcher/shared",
+        "//cmd/sourcegraph/osscmd",
         "//enterprise/cmd/embeddings/shared",
         "//enterprise/cmd/frontend/shared",
-        "//enterprise/cmd/sourcegraph/enterprisecmd",
         "//enterprise/cmd/symbols/shared",
         "//enterprise/cmd/worker/shared",
         "//internal/sanitycheck",
