+++ cmd/sourcegraph/osscmd/BUILD.bazel
@@ -1,11 +1,12 @@
 load("@io_bazel_rules_go//go:def.bzl", "go_library")
 
 go_library(
-    name = "executorcmd",
-    srcs = ["executorcmd.go"],
-    importpath = "github.com/sourcegraph/sourcegraph/enterprise/cmd/sourcegraph/enterprisecmd/executorcmd",
+    name = "osscmd",
+    srcs = ["osscmd.go"],
+    importpath = "github.com/sourcegraph/sourcegraph/cmd/sourcegraph/osscmd",
     visibility = ["//visibility:public"],
     deps = [
+        "//internal/authz",
         "//internal/service",
         "//internal/service/svcmain",
     ],
