+++ cmd/symbols/BUILD.bazel
@@ -9,7 +9,7 @@ go_library(
     importpath = "github.com/sourcegraph/sourcegraph/cmd/symbols",
     visibility = ["//visibility:private"],
     deps = [
-        "//cmd/sourcegraph-oss/osscmd",
+        "//cmd/sourcegraph/osscmd",
         "//cmd/symbols/shared",
         "//internal/sanitycheck",
     ],
