+++ enterprise/cmd/embeddings/BUILD.bazel
@@ -11,8 +11,8 @@ go_library(
     visibility = ["//visibility:private"],
     deps = [
         "//enterprise/cmd/embeddings/shared",
-        "//enterprise/cmd/sourcegraph/enterprisecmd",
         "//internal/sanitycheck",
+        "//internal/service/svcmain",
     ],
 )
