--- enterprise/cmd/sourcegraph/enterprisecmd/BUILD.bazel
@@ -1,12 +1,0 @@
-load("@io_bazel_rules_go//go:def.bzl", "go_library")
-
-go_library(
-    name = "enterprisecmd",
-    srcs = ["enterprisecmd.go"],
-    importpath = "github.com/sourcegraph/sourcegraph/enterprise/cmd/sourcegraph/enterprisecmd",
-    visibility = ["//visibility:public"],
-    deps = [
-        "//internal/service",
-        "//internal/service/svcmain",
-    ],
-)
