+++ enterprise/cmd/symbols/BUILD.bazel
@@ -10,9 +10,9 @@ go_library(
     importpath = "github.com/sourcegraph/sourcegraph/enterprise/cmd/symbols",
     visibility = ["//visibility:private"],
     deps = [
-        "//enterprise/cmd/sourcegraph/enterprisecmd",
         "//enterprise/cmd/symbols/shared",
         "//internal/sanitycheck",
+        "//internal/service/svcmain",
     ],
 )
