+++ enterprise/cmd/symbols/main.go
@@ -1,12 +1,14 @@
 package main
 
 import (
-	"github.com/sourcegraph/sourcegraph/enterprise/cmd/sourcegraph/enterprisecmd"
 	"github.com/sourcegraph/sourcegraph/enterprise/cmd/symbols/shared"
 	"github.com/sourcegraph/sourcegraph/internal/sanitycheck"
+	"github.com/sourcegraph/sourcegraph/internal/service/svcmain"
 )
+
+var config = svcmain.Config{}
 
 func main() {
 	sanitycheck.Pass()
-	enterprisecmd.SingleServiceMainEnterprise(shared.Service)
+	svcmain.SingleServiceMain(shared.Service, config)
 }
