+++ enterprise/dev/app/build-backend.sh
@@ -34,7 +34,7 @@ go_build() {
     -trimpath \
     -tags dist \
     -ldflags "$ldflags" \
-    ./enterprise/cmd/sourcegraph
+    ./cmd/sourcegraph
 }
 
 bazelrc() {
@@ -52,7 +52,7 @@ bazel_build() {
   local platform
   local bin_dir
   bazel_cmd="bazel"
-  bazel_target="//enterprise/cmd/sourcegraph:sourcegraph"
+  bazel_target="//cmd/sourcegraph:sourcegraph"
   bazel_opts="--stamp --workspace_status_command=./enterprise/dev/app/app-stamp-vars.sh"
   platform=$1
   bin_dir=$2
@@ -62,9 +62,9 @@ bazel_build() {
   fi
 
   # we need special flags and targets when cross compiling
-  # for more info see the BUILD.bazel file in enterprise/cmd/sourcegraph
+  # for more info see the BUILD.bazel file in cmd/sourcegraph
   if [[ ${CROSS_COMPILE_X86_64_MACOS:-0} == 1 ]]; then
-    bazel_target="//enterprise/cmd/sourcegraph:sourcegraph_x86_64_darwin"
+    bazel_target="//cmd/sourcegraph:sourcegraph_x86_64_darwin"
     # we don't use the incompat-zig-linux-amd64 bazel config here, since we need bazel to pick up the host cc
     bazel_opts="${bazel_opts} --platforms @zig_sdk//platform:darwin_amd64 --extra_toolchains @zig_sdk//toolchain:darwin_amd64"
   fi
@@ -73,7 +73,7 @@ bazel_build() {
   # shellcheck disable=SC2086
   ${bazel_cmd} build ${bazel_target} ${bazel_opts}
 
-  out=$(bazel cquery //enterprise/cmd/sourcegraph:sourcegraph --output=files)
+  out=$(bazel cquery //cmd/sourcegraph:sourcegraph --output=files)
   mkdir -p "${bin_dir}"
   chmod +x "${out}"
   cp -vf "${out}" "${bin_dir}/sourcegraph-backend-${platform}"
@@ -86,7 +86,6 @@ upload_artifacts() {
   target_path=$2
   buildkite-agent artifact upload "${target_path}"
 }
-
 
 # determine platform if it is not set
 PLATFORM=${PLATFORM:-"$(./enterprise/dev/app/detect-platform.sh)"}
